backend/sync-stream/requirements.txt
functions-framework==3.*
snowflake-connector-python==3.*
firebase-admin==6.*
google-cloud-secret-manager==2.*
